backend/requirements.txt
Flask==2.0.1
requests==2.26.0
firebase-admin==5.1.0
Pillow==9.1.0